self-voicing fixes
various changes that relate to self-voicing: (1) Adding character
definitions that appear the same visually when self-voicing is off, but
when sv is enabled, the computer will prompt the player to choose a
response. So far these characters cover up to the end of Act One. (2)
An image (the “Ten years later” card) didn’t have alt text. It has been
added. (3) It seemed strange that Stephanie’s name appeared before she
shared it with the player, so now she’s called “Librarian” during her
first line of dialogue.